+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - internal
   - remove imports from `pyo3` where covered by `pyo3::prelude::*`
+- `ryo3-tokio-rt`
+  - moved the `ryo3-tokio` runtime tools/utils into own crate `ryo3-tokio-rt`
+  - Use crate in `ryo3-reqwest`, `ryo3-tokio` and future `ryo3-tokio-websockets`
 
 ---